correct indexed grammar annotations and imports
