+++ pyformlang/indexed_grammar/duplication_rule.py
@@ -2,13 +2,12 @@
 A representation of a duplication rule, i.e. a rule that duplicates the stack
 """
 
-from typing import List, Set, Any
+from typing import List, Set, Hashable, Any
 
-from pyformlang.cfg import Variable, Terminal
-from pyformlang.cfg.utils import to_variable
-from pyformlang.cfg.cfg_object import CFGObject
+from pyformlang.cfg import CFGObject, Variable, Terminal
 
 from .reduced_rule import ReducedRule
+from ..objects.cfg_objects.utils import to_variable
 
 
 class DuplicationRule(ReducedRule):
@@ -25,6 +24,14 @@ class DuplicationRule(ReducedRule):
         The second non-terminal on the right of the rule (C here)
     """
 
+    def __init__(self,
+                 left_term: Hashable,
+                 right_term0: Hashable,
+                 right_term1: Hashable) -> None:
+        self._left_term = to_variable(left_term)
+        self._right_terms = (to_variable(right_term0),
+                             to_variable(right_term1))
+
     @property
     def f_parameter(self) -> Terminal:
         raise NotImplementedError
@@ -32,18 +39,6 @@ class DuplicationRule(ReducedRule):
     @property
     def production(self) -> Terminal:
         raise NotImplementedError
-
-    @property
-    def right_term(self) -> CFGObject:
-        raise NotImplementedError
-
-    def __init__(self,
-                 left_term: Any,
-                 right_term0: Any,
-                 right_term1: Any) -> None:
-        self._left_term = to_variable(left_term)
-        self._right_terms = (to_variable(right_term0),
-                             to_variable(right_term1))
 
     @property
     def left_term(self) -> Variable:
@@ -55,6 +50,10 @@ class DuplicationRule(ReducedRule):
             The left term of the rule
         """
         return self._left_term
+
+    @property
+    def right_term(self) -> CFGObject:
+        raise NotImplementedError
 
     @property
     def right_terms(self) -> List[CFGObject]:
@@ -89,13 +88,13 @@ class DuplicationRule(ReducedRule):
         """
         return set()
 
-    def __repr__(self) -> str:
-        """Gives a string representation of the rule, ignoring the sigmas"""
-        return f"{self._left_term} -> \
-            {self._right_terms[0]} {self._right_terms[1]}"
-
     def __eq__(self, other: Any) -> bool:
         if not isinstance(other, DuplicationRule):
             return False
         return other.left_term == self._left_term \
             and other.right_terms == self.right_terms
+
+    def __repr__(self) -> str:
+        """Gives a string representation of the rule, ignoring the sigmas"""
+        return f"{self._left_term} -> \
+            {self._right_terms[0]} {self._right_terms[1]}"
