+++ pyformlang/indexed_grammar/production_rule.py
@@ -2,13 +2,12 @@
 Represents a production rule, i.e. a rule that pushed on the stack
 """
 
-from typing import List, Set, Any
+from typing import List, Set, Hashable, Any
 
-from pyformlang.cfg import Variable, Terminal
-from pyformlang.cfg.utils import to_variable, to_terminal
-from pyformlang.cfg.cfg_object import CFGObject
+from pyformlang.cfg import CFGObject, Variable, Terminal
 
 from .reduced_rule import ReducedRule
+from ..objects.cfg_objects.utils import to_variable, to_terminal
 
 
 class ProductionRule(ReducedRule):
@@ -25,17 +24,17 @@ class ProductionRule(ReducedRule):
         The terminal used in the rule, "r" here
     """
 
-    @property
-    def f_parameter(self) -> Terminal:
-        raise NotImplementedError
-
     def __init__(self,
-                 left_term: Any,
-                 right_term: Any,
-                 production: Any) -> None:
+                 left_term: Hashable,
+                 right_term: Hashable,
+                 production: Hashable) -> None:
         self._left_term = to_variable(left_term)
         self._right_term = to_variable(right_term)
         self._production = to_terminal(production)
+
+    @property
+    def f_parameter(self) -> Terminal:
+        raise NotImplementedError
 
     @property
     def production(self) -> Terminal:
@@ -103,13 +102,13 @@ class ProductionRule(ReducedRule):
         """
         return {self._production}
 
-    def __repr__(self) -> str:
-        """Gets the string representation of the rule"""
-        return f"{self._left_term} -> {self._right_term} [ {self._production} ]"
-
     def __eq__(self, other: Any) -> bool:
         if not isinstance(other, ProductionRule):
             return False
         return other.left_term == self.left_term \
             and other.right_term == self.right_term \
             and other.production == self.production
+
+    def __repr__(self) -> str:
+        """Gets the string representation of the rule"""
+        return f"{self._left_term} -> {self._right_term} [ {self._production} ]"
